Move shared types to types.ts
