+++ packages/colorpickers/src/styled/ColorDialog/StyledDialogPreview.ts
@@ -7,7 +7,7 @@
 
 import styled from 'styled-components';
 import { getColor, retrieveComponentStyles, DEFAULT_THEME } from '@zendeskgarden/react-theming';
-import { IRGBColor } from '../../elements/ColorPicker/reducer';
+import { IRGBColor } from '../../utils/types';
 
 const COMPONENT_ID = 'colorpickers.colordialog.preview';
