+++ packages/colorpickers/src/styled/ColorPicker/StyledAlphaGradient.ts
@@ -7,7 +7,7 @@
 
 import styled from 'styled-components';
 import { retrieveComponentStyles, DEFAULT_THEME } from '@zendeskgarden/react-theming';
-import { IRGBColor } from '../../elements/ColorPicker/reducer';
+import { IRGBColor } from '../../utils/types';
 
 const COMPONENT_ID = 'colorpickers.colorpicker.alpha_gradient';
