+++ packages/colorpickers/src/styled/ColorPicker/StyledPreview.ts
@@ -7,7 +7,7 @@
 
 import styled, { ThemeProps, DefaultTheme } from 'styled-components';
 import { getColor, retrieveComponentStyles, DEFAULT_THEME } from '@zendeskgarden/react-theming';
-import { IRGBColor } from '../../elements/ColorPicker/reducer';
+import { IRGBColor } from '../../utils/types';
 
 const COMPONENT_ID = 'colorpickers.colorpicker.preview_box';
